Contact Information
This is the GitHub of Jonathan Tebo (B.S. Phy. '22). I was an undergrad in the J. Markert Lab at UT Austin from Sp20 through Fa22.

If you belong to the Markert Lab and would like to reach me, my e-mail address and cell phone number are written in the front cover of my lab notebook.
The contact information in my lab notebook is current as of December 2022. I intend to keep the same e-mail address and phone number forever.

The last versions of each of these files that Jonathan ever touched are copied to the lab computer on the equipment rack to the immediate left of the cryocooler.